+++ src/background/make-chromium-happy.ts
@@ -1,15 +1,18 @@
 import {MessageType} from '../utils/message';
 import type {Message} from '../definitions';
+import {isChromium} from '../utils/platform';
+
+declare const __MV3__: boolean;
 
 // This function exists to prevent Chrome from logging an error about
 // closed conduit. It just sends a dummy message in response to incomming message
 // to utilise open conduit. This response message is not even used on the other side.
 export function makeChromiumHappy() {
+    if (__MV3__ || !isChromium) {
+        return;
+    }
     chrome.runtime.onMessage.addListener((message: Message, _, sendResponse) => {
         if (![
-            // TabManager
-            MessageType.CS_FRAME_CONNECT,
-
             // Messenger
             MessageType.UI_GET_DATA,
             MessageType.UI_APPLY_DEV_DYNAMIC_THEME_FIXES,