+++ src/background/tab-manager.ts
@@ -2,10 +2,12 @@ import {canInjectScript} from '../background/utils/extension-api';
 import {createFileLoader} from './utils/network';
 import type {FetchRequestParameters} from './utils/network';
 import type {Message} from '../definitions';
-import {isFirefox, isMV3, isOpera, isThunderbird} from '../utils/platform';
+import {isFirefox, isOpera, isThunderbird} from '../utils/platform';
 import {MessageType} from '../utils/message';
 import {logWarn} from '../utils/log';
 import {StateManager} from './utils/state-manager';
+
+declare const __MV3__: boolean;
 
 async function queryTabs(query: chrome.tabs.QueryInfo) {
     return new Promise<chrome.tabs.Tab[]>((resolve) => {
@@ -21,7 +23,7 @@ interface ConnectionMessageOptions {
 interface TabManagerOptions {
     getConnectionMessage: (options: ConnectionMessageOptions) => Message | Promise<Message>;
     getTabMessage: (url: string, frameUrl: string) => Message;
-    onColorSchemeChange: ({isDark}: {isDark: boolean}) => void;
+    onColorSchemeChange: (isDark: boolean) => void;
 }
 
 interface FrameInfo {
@@ -63,6 +65,22 @@ export default class TabManager {
         this.tabs = {};
         this.getTabMessage = getTabMessage;
 
+        async function removeFrame(tabManager: TabManager, tabId: number, frameId: number){
+            await tabManager.stateManager.loadState();
+
+            if (frameId === 0) {
+                delete tabManager.tabs[tabId];
+            }
+
+            if (tabManager.tabs[tabId] && tabManager.tabs[tabId][frameId]) {
+                // We need to use delete here because Object.entries()
+                // in sendMessage() would enumerate undefined as well.
+                delete tabManager.tabs[tabId][frameId];
+            }
+
+            tabManager.stateManager.saveState();
+        }
+
         chrome.runtime.onMessage.addListener(async (message: Message, sender, sendResponse) => {
             function addFrame(tabs: {[tabId: number]: {[frameId: number]: FrameInfo}}, tabId: number, frameId: number, senderURL: string, timestamp: number) {
                 let frames: {[frameId: number]: FrameInfo};
@@ -81,6 +99,9 @@ export default class TabManager {
 
             switch (message.type) {
                 case MessageType.CS_FRAME_CONNECT: {
+                    if (__MV3__) {
+                        onColorSchemeChange(message.data.isDark);
+                    }
                     await this.stateManager.loadState();
                     const reply = (options: ConnectionMessageOptions) => {
                         const message = getConnectionMessage(options);
@@ -127,38 +148,27 @@ export default class TabManager {
                         frameURL: frameId === 0 ? null : senderURL,
                     });
                     this.stateManager.saveState();
-                    sendResponse({type: '¯\\_(ツ)_/¯'});
                     break;
                 }
-                case MessageType.CS_FRAME_FORGET: {
-                    await this.stateManager.loadState();
+                case MessageType.CS_FRAME_FORGET:
                     if (!sender.tab) {
                         logWarn('Unexpected message', message, sender);
                         break;
                     }
-                    const tabId = sender.tab.id;
-                    const frameId = sender.frameId;
-
-                    if (frameId === 0) {
-                        delete this.tabs[tabId];
-                    }
-
-                    if (this.tabs[tabId] && this.tabs[tabId][frameId]) {
-                        // We need to use delete here because Object.entries()
-                        // in sendMessage() would enumerate undefined as well.
-                        delete this.tabs[tabId][frameId];
-                    }
-                    this.stateManager.saveState();
+                    removeFrame(this, sender.tab.id, sender.frameId);
                     break;
-                }
-                case MessageType.CS_FRAME_FREEZE:
+                case MessageType.CS_FRAME_FREEZE: {
                     await this.stateManager.loadState();
                     const info = this.tabs[sender.tab.id][sender.frameId];
                     info.state = DocumentState.FROZEN;
                     info.url = null;
                     this.stateManager.saveState();
                     break;
+                }
                 case MessageType.CS_FRAME_RESUME: {
+                    if (__MV3__) {
+                        onColorSchemeChange(message.data.isDark);
+                    }
                     await this.stateManager.loadState();
                     const tabId = sender.tab.id;
                     const frameId = sender.frameId;
@@ -175,10 +185,9 @@ export default class TabManager {
                     this.stateManager.saveState();
                     break;
                 }
-                case MessageType.CS_DARK_THEME_DETECTED: {
+                case MessageType.CS_DARK_THEME_DETECTED:
                     this.tabs[sender.tab.id][sender.frameId].darkThemeDetected = true;
                     break;
-                }
 
                 case MessageType.CS_FETCH: {
                     // Using custom response due to Chrome and Firefox incompatibility
@@ -206,10 +215,11 @@ export default class TabManager {
                     break;
                 }
 
-                case MessageType.CS_COLOR_SCHEME_CHANGE: {
+                case MessageType.UI_COLOR_SCHEME_CHANGE:
+                    // fallthrough
+                case MessageType.CS_COLOR_SCHEME_CHANGE:
                     onColorSchemeChange(message.data);
                     break;
-                }
 
                 case MessageType.UI_SAVE_FILE: {
                     const {content, name} = message.data;
@@ -225,8 +235,13 @@ export default class TabManager {
                     chrome.tabs.sendMessage<Message>(activeTab.id, {type: MessageType.BG_EXPORT_CSS}, {frameId: 0});
                     break;
                 }
+
+                default:
+                    break;
             }
         });
+
+        chrome.tabs.onRemoved.addListener(async (tabId) => removeFrame(this, tabId, 0));
     }
 
     getTabURL(tab: chrome.tabs.Tab): string {
@@ -242,7 +257,7 @@ export default class TabManager {
             .filter((tab) => !Boolean(this.tabs[tab.id]))
             .forEach((tab) => {
                 if (!tab.discarded) {
-                    if (isMV3) {
+                    if (__MV3__) {
                         chrome.scripting.executeScript({
                             target: {
                                 tabId: tab.id,