+++ src/definitions.d.ts
@@ -1,5 +1,6 @@
 import type {ParsedColorSchemeConfig} from './utils/colorscheme-parser';
 import type {FilterMode} from './generators/css-filter';
+import type {MessageType} from './utils/message';
 
 export interface ExtensionData {
     isEnabled: boolean;
@@ -21,7 +22,7 @@ export interface ExtensionData {
 }
 
 export interface TabData {
-    type: string;
+    type: MessageType;
     data?: any;
 }
 
@@ -89,6 +90,12 @@ export interface ThemePreset {
     theme: Theme;
 }
 
+export interface Automation {
+    enabled: boolean;
+    mode: '' | 'time' | 'system' | 'location';
+    behavior: 'OnOff' | 'Scheme';
+}
+
 export interface UserSettings {
     enabled: boolean;
     fetchNews: boolean;
@@ -101,8 +108,7 @@ export interface UserSettings {
     changeBrowserTheme: boolean;
     syncSettings: boolean;
     syncSitesFixes: boolean;
-    automation: '' | 'time' | 'system' | 'location';
-    automationBehaviour: 'OnOff' | 'Scheme';
+    automation: Automation;
     time: TimeSettings;
     location: LocationSettings;
     previewNewDesign: boolean;
@@ -131,7 +137,7 @@ export interface TabInfo {
 }
 
 export interface Message {
-    type: string;
+    type: MessageType;
     data?: any;
     id?: number;
     error?: any;