+++ src/ui/popup/automation-page/index.tsx
@@ -1,11 +1,16 @@
 import {m} from 'malevic';
 import {getLocalMessage} from '../../../utils/locales';
 import {CheckBox, TimeRangePicker, TextBox, Button} from '../../controls';
-import type {ViewProps} from '../types';
 import DropDown from '../../controls/dropdown/index';
+import {MessageType} from '../../../utils/message';
+import type {Message} from '../../../definitions';
+import type {ViewProps} from '../types';
+import type {Automation} from 'definitions';
+
+declare const __MV3__: boolean;
 
 export default function AutomationPage(props: ViewProps) {
-    const isSystemAutomation = props.data.settings.automation === 'system';
+    const isSystemAutomation = props.data.settings.automation.mode === 'system';
     const locationSettings = props.data.settings.location;
     const values = {
         'latitude': {
@@ -64,14 +69,22 @@ export default function AutomationPage(props: ViewProps) {
         });
     }
 
+    function changeAutomationMode(mode: Automation['mode']) {
+        props.actions.changeSettings({automation: {...props.data.settings.automation, ...{mode, enabled: Boolean(mode)}}});
+    }
+
+    function changeAutomationBehavior(behavior: Automation['behavior']) {
+        props.actions.changeSettings({automation: {...props.data.settings.automation, ...{behavior}}});
+    }
+
     return (
         <div
             class={'automation-page'}
         >
             <div class="automation-page__line">
                 <CheckBox
-                    checked={props.data.settings.automation === 'time'}
-                    onchange={(e: {target: {checked: boolean}}) => props.actions.changeSettings({automation: e.target.checked ? 'time' : ''})}
+                    checked={props.data.settings.automation.mode === 'time'}
+                    onchange={(e: {target: {checked: boolean}}) => changeAutomationMode(e.target.checked ? 'time' : '')}
                 />
                 <TimeRangePicker
                     startTime={props.data.settings.time.activation}
@@ -84,8 +97,8 @@ export default function AutomationPage(props: ViewProps) {
             </p>
             <div class="automation-page__line automation-page__location">
                 <CheckBox
-                    checked={props.data.settings.automation === 'location'}
-                    onchange={(e: {target: {checked: boolean}}) => props.actions.changeSettings({automation: e.target.checked ? 'location' : ''})}
+                    checked={props.data.settings.automation.mode === 'location'}
+                    onchange={(e: {target: {checked: boolean}}) => changeAutomationMode(e.target.checked ? 'location' : '')}
                 />
                 <TextBox
                     class="automation-page__location__latitude"
@@ -121,14 +134,22 @@ export default function AutomationPage(props: ViewProps) {
                 <CheckBox
                     class="automation-page__system-dark-mode__checkbox"
                     checked={isSystemAutomation}
-                    onchange={(e: {target: {checked: boolean}}) => props.actions.changeSettings({automation: e.target.checked ? 'system' : ''})}
+                    onchange={(e: {target: {checked: boolean}}) => changeAutomationMode(e.target.checked ? 'system' : '')}
                 />
                 <Button
                     class={{
                         'automation-page__system-dark-mode__button': true,
                         'automation-page__system-dark-mode__button--active': isSystemAutomation,
                     }}
-                    onclick={() => props.actions.changeSettings({automation: isSystemAutomation ? '' : 'system'})}
+                    onclick={() => {
+                        if (__MV3__) {
+                            chrome.runtime.sendMessage<Message>({
+                                type: MessageType.UI_COLOR_SCHEME_CHANGE,
+                                data: {isDark: matchMedia('(prefers-color-scheme: dark)').matches}
+                            });
+                        }
+                        changeAutomationMode(isSystemAutomation ? '' : 'system');
+                    }}
                 >{getLocalMessage('system_dark_mode')}
                 </Button>
             </div>
@@ -136,8 +157,8 @@ export default function AutomationPage(props: ViewProps) {
                 {getLocalMessage('system_dark_mode_description')}
             </p>
             <DropDown
-                onChange={(selected: any) => props.actions.changeSettings({automationBehaviour: selected})}
-                selected={props.data.settings.automationBehaviour}
+                onChange={(selected: any) => changeAutomationBehavior(selected)}
+                selected={props.data.settings.automation.behavior}
                 options={[
                     {id: 'OnOff', content: 'Toggle on/off'},
                     {id: 'Scheme', content: 'Toggle dark/light'},