+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Unreleased
 
+- Don't manage styles that have a empty `href` attribute.
+- Use `navigator.UserAgentData` when possible.
+- Add a `<meta name="darkreader-lock">` detector, to disable Dark Reader when detected (only dynamic theme).
+- Fix filter theme for Firefox v102+
+- Correctly open static theme editor on Mobile.
+- Migrate automation settings to it's own object. WARNING: this can lead to data loss of currently used automation settings due to browser behavior.
+- Correctly handle empty URL's in `background-image` property.
+
+## 4.9.52 (June 28, 2022)
+
 - Correctly handle escaped characters in CSS `url(...)` values.
 
 ## 4.9.51 (May 27, 2022)